Add EventEmitters to decouple state game interface
In order to set up the game to handle different user interfaces
in the future, switching the game logic, save persist logic, and
interface display to communicate strictly through events. The event
strings are "constants" in the EventTypes.js file, so that can be
shared between the three module and ensure that the same strings
are used for the message names. Message display now listens for
the message event, and the interface has a map for the message key
and the value is either a string of a function that accepts meta
data about the message event. This will set the interface up in the
future for translations as well. The index.js file is solely
responsible for connecting up the listeners, and starting the first
call to the state to load the save file.