Add electron interface to the game
We add a new interface to our game by making the app compatible
with electron. That means we can no longer store the save file
in the same folder, because we may not have read/write permissions.
Therefore, we switch the file system state module we have to
use the app data folder for whatever platform is playing the game.

In the index.html, our renderer process is listening for events
from the new interface class we've created. We also moved a copy
of our prompt, sentiment, and message maps to the renderer process
and we'll want to evaluate if that is the correct place for those
of if they should exist as static properties on the interface
classes.

We also updated our state modules read method to return a promise
as well as emit events. This should give us more flexibility in
terms of how we deal with the initial read.

For fun, we're bringing in the Press Start 2P font from google
fonts and added some very simple styling to the index.html