Add finite state machine to game logic.
In order to set the game up to grow rapidly, we need a way
to represent the flow of the game as a graph of logical states.
This adds the XState library for state machines. The main game
logic is implemented as a single machine, and communication with
the interface happens the same as before, through events emitted
from the game event emitter. The events are emitted as either
entry or exit actions during state transitions.